backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.28.0
python-dotenv==1.0.0
pydantic==2.11.7
python-multipart==0.0.6
groq==1.1.1
httpx==0.28.1
requests==2.32.3
beautifulsoup4==4.12.2
pdfplumber==0.10.3
python-jose[cryptography]
websockets==15.0.1
apscheduler==3.10.4
sentence-transformers==3.0.1
playwright==1.52.0
